+++ frontend/app/src/main/res/navigation/nav_graph.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/mapFragment">
+    app:startDestination="@id/homeFragment">
 
     <fragment
         android:id="@+id/homeFragment"
@@ -31,7 +31,16 @@
         android:id="@+id/wishListFragment"
         android:name="com.triply.barrierfreetrip.WishlistFragment"
         android:label="fragment_wish_list"
-        tools:layout="@layout/fragment_wishlist" />
+        tools:layout="@layout/fragment_wishlist">
+        <action
+            android:id="@+id/action_wishListFragment_to_wisiListMapFragment"
+            app:destination="@id/wishListMapFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/wishListMapFragment"
+        android:name="com.triply.barrierfreetrip.WishlistMapFragment"
+        android:label="fragment_wish_list_map"
+        tools:layout="@layout/fragment_wishlist_map" />
     <fragment
         android:id="@+id/myPageFragment"
         android:name="com.triply.barrierfreetrip.MyPageFragment"
