+++ frontend/app/src/main/res/values/themes.xml
@@ -11,7 +11,7 @@
     </style>
 
     <style name="Theme.BarrierFreeTrip" parent="Base.Theme.BarrierFreeTrip" />
-    
+
     <!-- Custom font 적용 -->
 
     <style name="Theme.App.Starting" parent="Theme.SplashScreen">
